+++ pkgs/dart_mcp_server/CHANGELOG.md
@@ -42,6 +42,7 @@
 - Compact the output of the `analyze_files` tool to reduce token usage.
 - Start analysis server instances lazily, and time them out after 10 minutes of
   inactivity.
+- Enable `set_semantics` command for flutter_driver.
 
 # 0.1.2 (Dart SDK 3.11.0)